move io utils out to new file
