+++ lib/io.py
@@ -1,0 +1,20 @@
+import re
+import urllib2
+
+
+def open_data_file(url):
+    open_url = lambda d: urllib2.urlopen(d['url'])
+    protocols = {
+        'http':  open_url,
+        'https': open_url,
+        'file':  lambda d: d['path'],
+    }
+    regex = re.compile(
+        '^(?P<url>(?P<protocol>%s):\/\/(?P<path>.+))$' \
+        % '|'.join(protocols.keys())
+    )
+    match = re.match(regex, url)
+    if match:
+        args = match.groupdict()
+        return protocols[args['protocol']](args)
+    return None
